java/src/test/java/com/yourtion/leetcode/daily/m05/d18/SolutionTest.java
package com.yourtion.leetcode.daily.m05.d18;

import com.yourtion.leetcode.TestUtils;
import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("每日一题 - 20200518")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments("[2,3,-2,4]", 6),
                arguments("[-2,0,-1]", 0)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void maxProduct(String source, int res) {
        System.out.printf("runTest: %s , res: %d", source, res);
        int[] ss = TestUtils.stringToIntegerArray(source);
        Assertions.assertEquals(res, new Solution().maxProduct(ss));
    }
}